Extract Decal and Decal atlas from Storage class
