+++ drivers/gles3/rasterizer_gles3.h
@@ -39,6 +39,7 @@
 #include "servers/rendering/renderer_compositor.h"
 #include "storage/canvas_texture_storage.h"
 #include "storage/config.h"
+#include "storage/decal_atlas_storage.h"
 #include "storage/render_target_storage.h"
 #include "storage/texture_storage.h"
 
@@ -53,6 +54,7 @@ protected:
 	GLES3::Config config;
 	GLES3::CanvasTextureStorage canvas_texture_storage;
 	GLES3::TextureStorage texture_storage;
+	GLES3::DecalAtlasStorage decal_atlas_storage;
 	RasterizerStorageGLES3 storage;
 	RasterizerCanvasGLES3 canvas;
 	RasterizerSceneGLES3 scene;
@@ -62,6 +64,7 @@ protected:
 public:
 	RendererCanvasTextureStorage *get_canvas_texture_storage() { return &canvas_texture_storage; }
 	RendererTextureStorage *get_texture_storage() { return &texture_storage; }
+	RendererDecalAtlasStorage *get_decal_atlas_storage() { return &decal_atlas_storage; }
 	RendererStorage *get_storage() { return &storage; }
 	RendererCanvasRender *get_canvas() { return &canvas; }
 	RendererSceneRender *get_scene() { return &scene; }
