+++ servers/rendering/renderer_rd/renderer_compositor_rd.h
@@ -40,6 +40,7 @@
 #include "servers/rendering/renderer_rd/renderer_storage_rd.h"
 #include "servers/rendering/renderer_rd/shaders/blit.glsl.gen.h"
 #include "servers/rendering/renderer_rd/storage_rd/canvas_texture_storage.h"
+#include "servers/rendering/renderer_rd/storage_rd/decal_atlas_storage.h"
 #include "servers/rendering/renderer_rd/storage_rd/texture_storage.h"
 #include "servers/rendering/renderer_rd/uniform_set_cache_rd.h"
 
@@ -49,6 +50,7 @@ protected:
 	RendererCanvasRenderRD *canvas;
 	RendererRD::CanvasTextureStorage *canvas_texture_storage;
 	RendererRD::TextureStorage *texture_storage;
+	RendererRD::DecalAtlasStorage *decal_atlas_storage;
 	RendererStorageRD *storage;
 	RendererSceneRenderRD *scene;
 
@@ -92,8 +94,9 @@ protected:
 	static uint64_t frame;
 
 public:
-	RendererCanvasTextureStorage *get_canvas_texture_storage() { return canvas_texture_storage; };
-	RendererTextureStorage *get_texture_storage() { return texture_storage; };
+	RendererCanvasTextureStorage *get_canvas_texture_storage() { return canvas_texture_storage; }
+	RendererTextureStorage *get_texture_storage() { return texture_storage; }
+	RendererDecalAtlasStorage *get_decal_atlas_storage() { return decal_atlas_storage; }
 	RendererStorage *get_storage() { return storage; }
 	RendererCanvasRender *get_canvas() { return canvas; }
 	RendererSceneRender *get_scene() { return scene; }
