+++ servers/rendering/renderer_rd/renderer_storage_rd.h
@@ -205,50 +205,6 @@ private:
 	RID custom_rd_samplers[RS::CANVAS_ITEM_TEXTURE_FILTER_MAX][RS::CANVAS_ITEM_TEXTURE_REPEAT_MAX];
 	RID default_rd_storage_buffer;
 
-	/* DECAL ATLAS */
-
-	struct DecalAtlas {
-		struct Texture {
-			int panorama_to_dp_users;
-			int users;
-			Rect2 uv_rect;
-		};
-
-		struct SortItem {
-			RID texture;
-			Size2i pixel_size;
-			Size2i size;
-			Point2i pos;
-
-			bool operator<(const SortItem &p_item) const {
-				//sort larger to smaller
-				if (size.height == p_item.size.height) {
-					return size.width > p_item.size.width;
-				} else {
-					return size.height > p_item.size.height;
-				}
-			}
-		};
-
-		HashMap<RID, Texture> textures;
-		bool dirty = true;
-		int mipmaps = 5;
-
-		RID texture;
-		RID texture_srgb;
-		struct MipMap {
-			RID fb;
-			RID texture;
-			Size2i size;
-		};
-		Vector<MipMap> texture_mipmaps;
-
-		Size2i size;
-
-	} decal_atlas;
-
-	void _update_decal_atlas();
-
 	/* SHADER */
 
 	struct Material;
@@ -952,27 +908,6 @@ private:
 
 	mutable RID_Owner<ReflectionProbe, true> reflection_probe_owner;
 
-	/* DECAL */
-
-	struct Decal {
-		Vector3 extents = Vector3(1, 1, 1);
-		RID textures[RS::DECAL_TEXTURE_MAX];
-		float emission_energy = 1.0;
-		float albedo_mix = 1.0;
-		Color modulate = Color(1, 1, 1, 1);
-		uint32_t cull_mask = (1 << 20) - 1;
-		float upper_fade = 0.3;
-		float lower_fade = 0.3;
-		bool distance_fade = false;
-		float distance_fade_begin = 10;
-		float distance_fade_length = 1;
-		float normal_fade = 0.0;
-
-		Dependency dependency;
-	};
-
-	mutable RID_Owner<Decal, true> decal_owner;
-
 	/* VOXEL GI */
 
 	struct VoxelGI {
@@ -1190,17 +1125,6 @@ private:
 	EffectsRD *effects = nullptr;
 
 public:
-	RID decal_atlas_get_texture() const;
-	RID decal_atlas_get_texture_srgb() const;
-	_FORCE_INLINE_ Rect2 decal_atlas_get_texture_rect(RID p_texture) {
-		DecalAtlas::Texture *t = decal_atlas.textures.getptr(p_texture);
-		if (!t) {
-			return Rect2();
-		}
-
-		return t->uv_rect;
-	}
-
 	//internal usage
 
 	_FORCE_INLINE_ RID sampler_rd_get_default(RS::CanvasItemTextureFilter p_filter, RS::CanvasItemTextureRepeat p_repeat) {
@@ -1790,89 +1714,6 @@ public:
 
 	void base_update_dependency(RID p_base, DependencyTracker *p_instance);
 	void skeleton_update_dependency(RID p_skeleton, DependencyTracker *p_instance);
-
-	/* DECAL API */
-
-	RID decal_allocate();
-	void decal_initialize(RID p_decal);
-
-	virtual void decal_set_extents(RID p_decal, const Vector3 &p_extents);
-	virtual void decal_set_texture(RID p_decal, RS::DecalTexture p_type, RID p_texture);
-	virtual void decal_set_emission_energy(RID p_decal, float p_energy);
-	virtual void decal_set_albedo_mix(RID p_decal, float p_mix);
-	virtual void decal_set_modulate(RID p_decal, const Color &p_modulate);
-	virtual void decal_set_cull_mask(RID p_decal, uint32_t p_layers);
-	virtual void decal_set_distance_fade(RID p_decal, bool p_enabled, float p_begin, float p_length);
-	virtual void decal_set_fade(RID p_decal, float p_above, float p_below);
-	virtual void decal_set_normal_fade(RID p_decal, float p_fade);
-
-	void decal_atlas_mark_dirty_on_texture(RID p_texture);
-	void decal_atlas_remove_texture(RID p_texture);
-
-	virtual void texture_add_to_decal_atlas(RID p_texture, bool p_panorama_to_dp = false);
-	virtual void texture_remove_from_decal_atlas(RID p_texture, bool p_panorama_to_dp = false);
-
-	_FORCE_INLINE_ Vector3 decal_get_extents(RID p_decal) {
-		const Decal *decal = decal_owner.get_or_null(p_decal);
-		return decal->extents;
-	}
-
-	_FORCE_INLINE_ RID decal_get_texture(RID p_decal, RS::DecalTexture p_texture) {
-		const Decal *decal = decal_owner.get_or_null(p_decal);
-		return decal->textures[p_texture];
-	}
-
-	_FORCE_INLINE_ Color decal_get_modulate(RID p_decal) {
-		const Decal *decal = decal_owner.get_or_null(p_decal);
-		return decal->modulate;
-	}
-
-	_FORCE_INLINE_ float decal_get_emission_energy(RID p_decal) {
-		const Decal *decal = decal_owner.get_or_null(p_decal);
-		return decal->emission_energy;
-	}
-
-	_FORCE_INLINE_ float decal_get_albedo_mix(RID p_decal) {
-		const Decal *decal = decal_owner.get_or_null(p_decal);
-		return decal->albedo_mix;
-	}
-
-	_FORCE_INLINE_ uint32_t decal_get_cull_mask(RID p_decal) {
-		const Decal *decal = decal_owner.get_or_null(p_decal);
-		return decal->cull_mask;
-	}
-
-	_FORCE_INLINE_ float decal_get_upper_fade(RID p_decal) {
-		const Decal *decal = decal_owner.get_or_null(p_decal);
-		return decal->upper_fade;
-	}
-
-	_FORCE_INLINE_ float decal_get_lower_fade(RID p_decal) {
-		const Decal *decal = decal_owner.get_or_null(p_decal);
-		return decal->lower_fade;
-	}
-
-	_FORCE_INLINE_ float decal_get_normal_fade(RID p_decal) {
-		const Decal *decal = decal_owner.get_or_null(p_decal);
-		return decal->normal_fade;
-	}
-
-	_FORCE_INLINE_ bool decal_is_distance_fade_enabled(RID p_decal) {
-		const Decal *decal = decal_owner.get_or_null(p_decal);
-		return decal->distance_fade;
-	}
-
-	_FORCE_INLINE_ float decal_get_distance_fade_begin(RID p_decal) {
-		const Decal *decal = decal_owner.get_or_null(p_decal);
-		return decal->distance_fade_begin;
-	}
-
-	_FORCE_INLINE_ float decal_get_distance_fade_length(RID p_decal) {
-		const Decal *decal = decal_owner.get_or_null(p_decal);
-		return decal->distance_fade_length;
-	}
-
-	virtual AABB decal_get_aabb(RID p_decal) const;
 
 	/* VOXEL GI API */
